Create Nnaji -miracle-mmesoma -2024704132
This python program implements a 
grading system using ìf and elìf 
Statements. It takes a student's score 
as input and print the corresponding 
grade based on the following scale:

-70—100: Grade A
-60—69: Grade B
-50—59: Grade C
-45—49: Grade D
-40—44: Grade E
-20—39: Grade F

For each valid score, the program also 
prints "YES" to confirm the grading 
was successful. If the score is outside 
the valid range, it prints "invalid score" and "NO".

 Created by : Nnaji miracle mmesoma 
Reg. No: 2024704132.